[jwies/review-pi-harness-seam] review: split the per-agent dispatch machinery out of dispatch.ts
Both dispatch.ts (1011) and dispatch.test.ts (1066) crossed eslint's
1000-line max-lines cap once the retry/refusal work landed, so Lints and
the push-test job are red on this branch. Main is already at 950/981, so
a disable comment would just move the wall one commit further out.

Split along the seam the file already has: dispatch-calls.ts owns calling
ONE agent (the model seam types, the per-agent report entry, the
usage-policy refusal fallback, and the single malformed-output retry) and
knows nothing about rosters, phases, gating, or note lines. runDispatch
owns the run and gets the two closures from createAgentDispatcher; the
moved code is byte-identical apart from perAgent.push becoming an
injected report callback. dispatch.ts re-exports the seam types, so
dispatch-runner-pi and its test import exactly what they did before.

The malformed-output/retry describe block moves to dispatch-retry.test.ts
alongside it, matching the sibling split of dispatch-gate-hardening and
dispatch-trial-followups. Also drops a section banner in dispatch.ts left
empty when loadAgents moved to dispatch-agents.

756/351/813/373 lines; 115 tests over the four files, same count as before.
